Move license methods to util class instead of apiimpl
